make openmdao optional
